+++ HousingSearchApi/V2/Gateways/Interfaces/ISearchGateway.cs
@@ -1,9 +1,9 @@
 using System.Collections.Generic;
 using System.Threading.Tasks;
 
-namespace HousingSearchApi.V1.Gateways.Interfaces;
+namespace HousingSearchApi.V2.Gateways.Interfaces;
 
-public interface IGetGateway
+public interface ISearchGateway
 {
     Task<IReadOnlyCollection<object>> FreeSearch(string indexName, string searchText);
 }
