+++ HousingSearchApi/V2/UseCase/Interfaces/ISearchUseCase.cs
@@ -1,0 +1,10 @@
+using System.Collections.Generic;
+using System.Threading.Tasks;
+
+namespace HousingSearchApi.V2.UseCase.Interfaces;
+
+public interface ISearchUseCase
+{
+    Task<IReadOnlyCollection<object>> ExecuteAsync(string indexName, string searchText);
+}
+
